refactor: move handlers to their own package
