+++ src/test/java/uk/gov/hmcts/reform/wacaseeventhandler/services/WarningTaskHandlerTest.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.wacaseeventhandler.services;
 
 import org.junit.jupiter.api.Test;
 import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.EventInformation;
+import uk.gov.hmcts.reform.wacaseeventhandler.handlers.WarningTaskHandler;
 
 import static org.assertj.core.api.Assertions.assertThat;
